produce require optimizations/fixes
