+++ produce/lib/produce/developer_center.rb
@@ -1,4 +1,5 @@
 require 'spaceship'
+require_relative 'module'
 
 module Produce
   class DeveloperCenter
