+++ produce/lib/produce/group.rb
@@ -1,4 +1,5 @@
 require 'spaceship'
+require_relative 'module'
 
 module Produce
   class Group
