+++ produce/lib/produce/itunes_connect.rb
@@ -1,4 +1,6 @@
-require 'spaceship'
+require 'spaceship/tunes/tunes'
+require_relative 'module'
+require_relative 'available_default_languages'
 
 module Produce
   class ItunesConnect
@@ -53,13 +55,13 @@ module Produce
         UI.success "Successfully created new app '#{Produce.config[:app_name]}' on iTunes Connect with ID #{application.apple_id}"
       end
 
-      return Spaceship::Application.find(@full_bundle_identifier, mac: Produce.config[:platform] == "osx").apple_id
+      return Spaceship::Tunes::Application.find(@full_bundle_identifier, mac: Produce.config[:platform] == "osx").apple_id
     end
 
     private
 
     def fetch_application
-      Spaceship::Application.find(@full_bundle_identifier)
+      Spaceship::Tunes::Application.find(@full_bundle_identifier)
     end
 
     def wildcard_bundle?
